Mesh Networking in Android by using Wi-Fi Direct technology
MANET networks are those in which we don´t need a centered access point
so as to stablish a connection between two mobile points within the
network. Indeed, we are able to stablish peer to peer connection
without that centered access point, being able to come up with utterly
useful applications. In this project, design and development of an
application using MANET network is figured out.